+++ querier/queries/drop.py
@@ -2,11 +2,12 @@
 #
 # License: BSD 3
 
-
-from ..utils import memoize
+import pandas as pd
+import polars as pl
+from ..utils import polars_to_pandas, pandas_to_polars
 
 # creates a copy
-@memoize
+
 def drop(df, req=None):
     """ Drop columns.
        
@@ -22,7 +23,15 @@ def drop(df, req=None):
     """
 
     try:
-        return df.drop(req.replace(" ", "").split(","), axis=1)
+        if isinstance(df, pl.DataFrame):
+            df = polars_to_pandas(df)
+
+        result = df.drop(req.replace(" ", "").split(","), axis=1)
+
+        if isinstance(df, pl.DataFrame):
+            return pandas_to_polars(result)
+
+        return result
     except:
         raise ValueError(
             "request must contain df" "s column names (comma-separated)"