+++ querier/queries/request.py
@@ -4,9 +4,11 @@
 
 
 import pandas as pd
-from sqlalchemy import create_engine
+import polars as pl
+from sqlalchemy import create_engine, text 
 from ..utils import parse_cols_request
-from ..utils import memoize
+from ..utils import polars_to_pandas, pandas_to_polars
+
 
 # request(df, "SELECT tip, smoker, day FROM df WHERE tip > 25")
 # request(df, "SELECT tip, smoker, day, size FROM df WHERE (tip > 2) and (size > 4)")
@@ -17,8 +19,8 @@ from ..utils import memoize
 # request(df, "SELECT AVG(tip) as avg_tip, size, smoker FROM df WHERE tip > 20 GROUP BY size")
 # request(df, "SELECT avg(tip) as avg_tip, AVG(size) as   avg_size, smoker FROM df WHERE tip > 20 GROUP BY size, tip")
 # request(df, "SELECT SUM(tip), smoker FROM df GROUP BY smoker having tip > 1.5")
-@memoize
-def request(df, req=None, **kwargs):
+
+def request(df, req, **kwargs):
     """ SQL request on a data frame.
        
     Args:           
@@ -35,10 +37,6 @@ def request(df, req=None, **kwargs):
         https://github.com/thierrymoudiki/querier/tree/master/querier/demo
        
     """
-
-    if req is None:  # useless tho...
-
-        return df
 
     # if req is not None:
     assert (
@@ -57,11 +55,13 @@ def request(df, req=None, **kwargs):
 
         engine = create_engine("sqlite://", echo=False)
         df.to_sql("df", con=engine, **kwargs)
-        req_res = engine.execute(req.upper()).fetchall()
+        with engine.connect() as conn:
+            req_res = conn.execute(text(req.upper())).fetchall()
         col_names = parse_cols_request(req)
-        df_res = pd.DataFrame(req_res, columns=col_names)
-
-        return df_res
+        result = pd.DataFrame(req_res, columns=col_names)
+        if isinstance(df, pl.DataFrame):
+            return(pandas_to_polars(result))
+        return result
 
     except: