+++ querier/utils/__init__.py
@@ -1,4 +1,5 @@
 from .deepcopy import deepcopy
+from .convert import pandas_to_polars, polars_to_pandas
 from .memoize import memoize
 from .parse_request import (
     parse_request,
@@ -9,6 +10,8 @@ from .parse_request import (
 
 __all__ = [
     deepcopy,
+    pandas_to_polars,
+    polars_to_pandas,
     memoize,
     parse_request,
     parse_update_request,