+++ CHANGES.md
@@ -1,3 +1,10 @@
+# version 0.5.0
+
+- Add class `Dowloader` to get data from the R-universe API
+- Support for polars DataFrame
+- More technical: execute SQL commands after converting explicitly
+  to `text`
+
 # version 0.3.0
 
 - import data from sql and csv